requirements-test.txt
-r requirements.txt
openai-whisper
rich
diff-match-patch
playwright
pytest
